refactor adapters and utils
